% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-txflow_azureblobs.R
\docType{class}
\name{.txflow_azureblobs-class}
\alias{.txflow_azureblobs-class}
\alias{.txflow_azureblobs}
\title{An internal reference class representing txflow store using Azure block blob store}
\description{
A collection of internal methods to manage data file and blobs stored using
Azure Block Blob storage.

The storage relies on a Microsoft Azure Storage Account with each data
repository represented as a container and data files and blobs stored in the
root of the container. The container is named as the repository name without
a prefix making a shared storage account not practical.

The storage uses the following configuration options.

\itemize{
\item \code{TXFLOW.STORE} equal to \code{AZUREBLOBS}, case insensitive
\item \code{AZUREBLOBS.URL} specifying the URL to the blob storage
}

Access tokens for storage integration uses Microsoft Azure OAuth
authentication. The Azure REST API calls rely on authorization bearer token.

\itemize{
\item \code{AZURE.OAUTH.URL} is the Microsoft Azure OAuth URL including the full
URL path that should include the tenant ID
\item \code{AZURE.OAUTH.CLIENTID} the client ID associated with the storage account
\item \code{AZURE.OAUTH.CLIENTSECRET} the client secret associated with the storage
account
}

The scope of the requested access token is \verb{https://storage.azure.com}.

Note that the configuration relies on \link[cxapp]{cxapp_config} that supports
environmental variables and secret vaults.
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal storage of configuration attributes}
}}

\section{Methods}{

\describe{
\item{\code{commit(x, repository = NULL, overwrite = FALSE)}}{Commit files to repository}

\item{\code{exists(x)}}{Resource exists}

\item{\code{getreference(x)}}{Get resource reference}

\item{\code{getresource(x)}}{Get resource}

\item{\code{initialize()}}{Initialize}

\item{\code{is.locked(x)}}{Assert if snapshot is locked}

\item{\code{lock(x)}}{Lock snapshot}

\item{\code{repositories()}}{List repositories}

\item{\code{repository(x, create = FALSE)}}{Get a reference to a repository}

\item{\code{snapshot(x, file = NULL)}}{Get repository snapshot}

\item{\code{snapshots(x)}}{List snapshots of a repository}

\item{\code{unlock(x)}}{Unlock snapshot}
}}

