% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-txflow_defaultsnapshot.R
\name{.txflow_defaultsnapshot}
\alias{.txflow_defaultsnapshot}
\title{Internal function to return a default snapshot specification}
\usage{
.txflow_defaultsnapshot(x)
}
\arguments{
\item{x}{Snapshot}
}
\value{
A named list representing a defualt specification
}
\description{
Internal function to return a default snapshot specification
}
\keyword{internal}
