% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-txflow_fsstore.R
\docType{class}
\name{.txflow_fsstore-class}
\alias{.txflow_fsstore-class}
\alias{.txflow_fsstore}
\title{A reference class representing txflow store using the local file system}
\description{
A reference class representing txflow store using the local file system
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal storage of configuration attributes}
}}

\section{Methods}{

\describe{
\item{\code{commit(x, repository = NULL, overwrite = FALSE)}}{Commit files to repository}

\item{\code{exists(x)}}{Resource exists}

\item{\code{getreference(x)}}{Get resource reference}

\item{\code{getresource(x)}}{Get resource}

\item{\code{initialize()}}{Initialize}

\item{\code{is.locked(x)}}{Assert if snapshot is locked}

\item{\code{lock(x)}}{Lock snapshot}

\item{\code{repositories()}}{List repositories}

\item{\code{repository(x, create = FALSE)}}{Get a reference to a repository}

\item{\code{snapshot(x, file = NULL)}}{Get repository snapshot}

\item{\code{snapshots(repository)}}{List snapshots of a repository}

\item{\code{unlock(x)}}{Unlock snapshot}
}}

