% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start.R
\name{start}
\alias{start}
\title{Start txflow service}
\usage{
start(port = 12345)
}
\arguments{
\item{port}{Service port}
}
\description{
A standard configurable start for plumber.

The function searches for \code{plumber.R} in the following directories
\itemize{
\item current working directory
\item directory \code{inst/ws} in the current working directory (development mode)
\item directory \code{ws} in the current working directory (installed)
\item directory \code{ws} in the first occurrence of \code{txflow.service} in the library tree
}

Note that the above configuration uses \link[cxapp]{cxapp_config} and supports
both environmental variables and key vaults.
}
