% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_addfile.R
\name{txflow_addfile}
\alias{txflow_addfile}
\title{Add a file to a draft snapshot}
\usage{
txflow_addfile(x, work = NULL, attrs = NULL)
}
\arguments{
\item{x}{File}

\item{work}{Work area}

\item{attrs}{File attributes}
}
\value{
File details
}
\description{
Add a file to a draft snapshot
}
