% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_addreference.R
\name{txflow_addreference}
\alias{txflow_addreference}
\title{Utility function to add reference to an existing repository file to a work area}
\usage{
txflow_addreference(x, work = NULL, name = NULL)
}
\arguments{
\item{x}{Resource}

\item{work}{Work area}

\item{name}{Name for resource}
}
\value{
Resource specification
}
\description{
A reference can only be added to a work area that contains a snapshot
specification. All added references must be from the same repository as
the snapshot.

The resource can be specified simply as \verb{<resource>} as the repository is
known. If the resource is specified as either a repository resource
\verb{<repository>/<resource>} or explicitly as a snapshot reference
\verb{<repository>/<snapshot>/<resource>}, the repository must match the
repository in the snapshot specification.
}
