% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_commit.R
\name{txflow_commit}
\alias{txflow_commit}
\title{Utility function to commit changes to a snapshot to the repository}
\usage{
txflow_commit(x, snapshot = NULL, as.actor = NULL)
}
\arguments{
\item{x}{Work area}

\item{snapshot}{Snapshot}

\item{as.actor}{User or service principal performing the commit}
}
\value{
Logical \code{TRUE} if commit succeeded
}
\description{
The snapshot \code{snapshot} is referenced as \verb{<repository>/<snapshot>}.
}
