% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_dropfile.R
\name{txflow_dropfile}
\alias{txflow_dropfile}
\title{Utility function remove a file from a snapshot}
\usage{
txflow_dropfile(x, work = NULL)
}
\arguments{
\item{x}{Reference to resource to remove}

\item{work}{Work area}
}
\value{
Invisible \code{TRUE}
}
\description{
The resource reference \code{x} can either be the resource name or blob.
}
