% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_dropworkarea.R
\name{txflow_dropworkarea}
\alias{txflow_dropworkarea}
\title{Utility function to drop a temporary work area}
\usage{
txflow_dropworkarea(x)
}
\arguments{
\item{x}{Work area reference}
}
\value{
Invisible logical
}
\description{
A work area is a temporary directory that is used as a \emph{working copy}.
}
