% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_getresource.R
\name{txflow_getresource}
\alias{txflow_getresource}
\title{Utility function to retrieve a file associated with a repository snapshot}
\usage{
txflow_getresource(x)
}
\arguments{
\item{x}{Snapshot file or resource}
}
\value{
Path to the cached file or resource
}
\description{
A file or resource is an item that is stored in a data repository and
associated with a snapshot.

The file or resource \code{x} is specified in the format
\verb{<repository>/<snapshot>/<file or resource>}.

The utility uses the cxapp application cache \link[cxapp]{cxapp_applicationcache}
as a temporary storage location for the resource or file retrieved from the
back end storage solution.
}
