% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_islocked.R
\name{txflow_islocked}
\alias{txflow_islocked}
\title{Utility function to assert if a snapshot is locked}
\usage{
txflow_islocked(x)
}
\arguments{
\item{x}{Resource}
}
\value{
Invisible logical
}
\description{
Asserts if resource \code{x} is locked

Returns \code{TRUE} is if the resource is locked. \code{FALSE} otherwise.
}
