% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_listrepositories.R
\name{txflow_listrepositories}
\alias{txflow_listrepositories}
\title{List data repositories}
\usage{
txflow_listrepositories()
}
\value{
Vector of repositories
}
\description{
List data repositories
}
