% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_listsnapshots.R
\name{txflow_listsnapshots}
\alias{txflow_listsnapshots}
\title{List snapshots in a repository}
\usage{
txflow_listsnapshots(x)
}
\arguments{
\item{x}{Repository}
}
\value{
A vector of snapshot names
}
\description{
List snapshots in a repository
}
