% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_repository.R
\name{txflow_repository}
\alias{txflow_repository}
\title{Create or get reference to an existing repository}
\usage{
txflow_repository(x, as.actor = NULL)
}
\arguments{
\item{x}{Repository name}

\item{as.actor}{Principal creating the repository}
}
\value{
Repository name
}
\description{
Create or get reference to an existing repository
}
