% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_snapshot.R
\name{txflow_snapshot}
\alias{txflow_snapshot}
\title{Utility function to get a snapshot specification}
\usage{
txflow_snapshot(x, as.actor = NULL, audited = FALSE)
}
\arguments{
\item{x}{Snapshot}

\item{as.actor}{User of service requesting snapshot specification}

\item{audited}{Enable audit trail for snapshot specification request}
}
\value{
Snapshot specification as a list
}
\description{
Returns a snapshot definition

The snapshot \code{x} is referenced as \verb{<repository>/<snapshot>}.
}
