% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_store.R
\name{txflow_store}
\alias{txflow_store}
\title{Utility function to connect the txflow storage configuration}
\usage{
txflow_store()
}
\value{
A tx flow storage object
}
\description{
Utility function to connect the txflow storage configuration
}
