% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_validname.R
\name{txflow_validname}
\alias{txflow_validname}
\title{Function to validate Flow references names}
\usage{
txflow_validname(x, context = NULL)
}
\arguments{
\item{x}{Character string}

\item{context}{Force context}
}
\value{
Invisible logical
}
\description{
Flow repository names are restricted by the different service that actually
store the repository.

A value delimited by a foward slash \code{/} is assumed to represent a snapshot
reference. A value without the slash delimited represents a repsitory.

If name \code{x} is valid, \code{TRUE} is returned. Otherwise, \code{FALSE}.

A repository name is restricted to the following conditions

\itemize{
\item Consists of letters a-z, digits 0-9 and punctuation dash \code{-}
\item Starts and ends with a letter or number
\item All dashes \code{-} must be preceded and followed by letter or digit
\item Length is between 3 and 63 characters
}

A snapshot name is restricted to the following conditions

\itemize{
\item Consists of letters a-z, digits 0-9 and punctuation dash \code{-} and underscore \verb{_}
\item Starts and ends with a letter or number
\item All dashes \code{-} and underscores \verb{_} must be preceded and followed by letter or digit
\item Length is between 3 and 256 characters
}

Reference names are case insensitive
}
