% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_validtype.R
\name{txflow_validtype}
\alias{txflow_validtype}
\title{Utility function to assert if a data file or blob type is valid}
\usage{
txflow_validtype(x)
}
\arguments{
\item{x}{The type}
}
\value{
Invisible logical
}
\description{
Asserts if \code{x} is a valid type reference.

A type is a character string
\itemize{
\item case insensitive
\item containing the letters A-Z, digits 0-9 and punctuations period
\code{.} and underscore \verb{_}
\item starts with a character or digit
\item consist of at least 1 character and maximum of 64
}

\code{type} with prefix \verb{custom:} is assumed a non-standard bespoke type.

Supported standard types are
\itemize{
\item \code{datafile}
\item \code{datafile.sas}
\item \code{datafile.rds}
\item \code{datafile.rda}
\item \code{datafile.csv}
\item \code{datafile.xlsx}
\item \code{datafile.text}
}
}
