% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txflow_workarea.R
\name{txflow_workarea}
\alias{txflow_workarea}
\title{Utility function to create or get references to a temporary work area}
\usage{
txflow_workarea(work = NULL, snapshot = NULL)
}
\arguments{
\item{work}{Work area reference}

\item{snapshot}{Snapshot to clone}
}
\value{
Invisible work area path
}
\description{
A work area is a temporary directory that is used as a \emph{working copy}.

If \code{work} is not specified or \code{NULL}, a new work area is created

If \code{snapshot} is specified and it exists, the snapshot definition file
is staged in the work area.
}
