% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal-txflow_fsstore.R
\docType{class}
\name{.txflow_fsstore-class}
\alias{.txflow_fsstore-class}
\alias{.txflow_fsstore}
\title{A reference class representing txflow store using the local file system}
\description{
A collection of internal methods to manage data file and blobs stored using
the local file system.

The storage relies on the local file sysmte with each data repository
represented as a directory and data files and blobs stored in the
root of the repository directory. The directory is named as the repository
name without a prefix making using the repository parent directory a
shared storage area not practical.

The storage uses the following configuration options.

\itemize{
\item \code{TXFLOW.STORE} equal to \code{LOCAL}, case insensitive
\item \code{TXFLOW.DATA} specifying the parent directory for repositories
}
}
\section{Fields}{

\describe{
\item{\code{.attr}}{Internal storage of configuration attributes}
}}

\section{Methods}{

\describe{
\item{\code{commit(x, repository = NULL, overwrite = FALSE)}}{Commit files to repository}

\item{\code{exists(x)}}{Resource exists}

\item{\code{getreference(x)}}{Get resource reference}

\item{\code{getresource(x)}}{Get resource}

\item{\code{initialize()}}{Initialize}

\item{\code{is.locked(x)}}{Assert if snapshot is locked}

\item{\code{lock(x)}}{Lock snapshot}

\item{\code{repositories()}}{List repositories}

\item{\code{repository(x, create = FALSE)}}{Get a reference to a repository}

\item{\code{snapshot(x, file = NULL)}}{Get repository snapshot}

\item{\code{snapshots(repository)}}{List snapshots of a repository}

\item{\code{unlock(x)}}{Unlock snapshot}
}}

